getAllUsersForAdmin method moved from UserController to ManageUsersController
